.js to .ts(tsx) for services / store
