--- src/services/localQuery.js
@@ -1,8 +1,0 @@
-import { nanoid } from "nanoid";
-import { _QUERY_ITEM } from "./consts";
-
-import { getStorageItem, setStorageItem, cleanup } from "./localService";
-
-const localQuery = () => {};
-
-export default localQuery;
