+++ src/services/localQuery.ts
@@ -1,0 +1,8 @@
+// import { nanoid } from "nanoid";
+// import { _QUERY_ITEM } from "./consts";
+
+// import { getStorageItem, setStorageItem, cleanup } from "./localService";
+
+const localQuery = () => {};
+
+export default localQuery;
